shift sql over
